Rename engine_scripts/lib, lib is excluded from linting/prettier
