+++ testing/visual-regression/engine_scripts/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  env: {
+    browser: true,
+  },
+  rules: {
+    'no-console': 'off',
+  },
+};
